Migrated DX11 renderer, implemented device factory functions
