+++ Code/Engine/RendererDX11/CommandEncoder/RenderCommandEncoderDX11.h
@@ -1,0 +1,68 @@
+
+#pragma once
+
+#include <RendererDX11/CommandEncoder/CommandEncoderDX11.h>
+#include <RendererFoundation/CommandEncoder/RenderCommandEncoder.h>
+#include <RendererFoundation/Resources/RenderTargetSetup.h>
+
+class EZ_RENDERERDX11_DLL ezGALRenderCommandEncoderDX11 : public ezGALCommandEncoderDX11<ezGALRenderCommandEncoder>
+{
+protected:
+  friend class ezGALDeviceDX11;
+  friend class ezGALPassDX11;
+  friend class ezMemoryUtils;
+
+  using SUPER = ezGALCommandEncoderDX11<ezGALRenderCommandEncoder>;
+
+  ezGALRenderCommandEncoderDX11(ezGALDevice& device);
+  virtual ~ezGALRenderCommandEncoderDX11();
+
+  void BeginEncode(const ezGALRenderingSetup& renderingSetup);
+  void EndEncode();
+
+  // Draw functions
+
+  virtual void ClearPlatform(const ezColor& ClearColor, ezUInt32 uiRenderTargetClearMask, bool bClearDepth, bool bClearStencil, float fDepthClear, ezUInt8 uiStencilClear) override;
+
+  virtual void DrawPlatform(ezUInt32 uiVertexCount, ezUInt32 uiStartVertex) override;
+  virtual void DrawIndexedPlatform(ezUInt32 uiIndexCount, ezUInt32 uiStartIndex) override;
+  virtual void DrawIndexedInstancedPlatform(ezUInt32 uiIndexCountPerInstance, ezUInt32 uiInstanceCount, ezUInt32 uiStartIndex) override;
+  virtual void DrawIndexedInstancedIndirectPlatform(const ezGALBuffer* pIndirectArgumentBuffer, ezUInt32 uiArgumentOffsetInBytes) override;
+  virtual void DrawInstancedPlatform(ezUInt32 uiVertexCountPerInstance, ezUInt32 uiInstanceCount, ezUInt32 uiStartVertex) override;
+  virtual void DrawInstancedIndirectPlatform(const ezGALBuffer* pIndirectArgumentBuffer, ezUInt32 uiArgumentOffsetInBytes) override;
+  virtual void DrawAutoPlatform() override;
+
+  virtual void BeginStreamOutPlatform() override;
+  virtual void EndStreamOutPlatform() override;
+
+  // State functions
+
+  virtual void SetIndexBufferPlatform(const ezGALBuffer* pIndexBuffer) override;
+  virtual void SetVertexBufferPlatform(ezUInt32 uiSlot, const ezGALBuffer* pVertexBuffer) override;
+  virtual void SetVertexDeclarationPlatform(const ezGALVertexDeclaration* pVertexDeclaration) override;
+  virtual void SetPrimitiveTopologyPlatform(ezGALPrimitiveTopology::Enum Topology) override;
+
+  virtual void SetBlendStatePlatform(const ezGALBlendState* pBlendState, const ezColor& BlendFactor, ezUInt32 uiSampleMask) override;
+  virtual void SetDepthStencilStatePlatform(const ezGALDepthStencilState* pDepthStencilState, ezUInt8 uiStencilRefValue) override;
+  virtual void SetRasterizerStatePlatform(const ezGALRasterizerState* pRasterizerState) override;
+
+  virtual void SetViewportPlatform(const ezRectFloat& rect, float fMinDepth, float fMaxDepth) override;
+  virtual void SetScissorRectPlatform(const ezRectU32& rect) override;
+
+  virtual void SetStreamOutBufferPlatform(ezUInt32 uiSlot, const ezGALBuffer* pBuffer, ezUInt32 uiOffset) override;
+
+  virtual void FlushDeferredStateChanges() override;
+
+private:
+  // Bound objects for deferred state flushes
+  ezGALRenderTargetSetup m_RenderTargetSetup;
+  ID3D11RenderTargetView* m_pBoundRenderTargets[EZ_GAL_MAX_RENDERTARGET_COUNT] = {};
+  ezUInt32 m_uiBoundRenderTargetCount = 0;
+  ID3D11DepthStencilView* m_pBoundDepthStencilTarget = nullptr;
+
+  ID3D11Buffer* m_pBoundVertexBuffers[EZ_GAL_MAX_VERTEX_BUFFER_COUNT] = {};
+  ezGAL::ModifiedRange m_BoundVertexBuffersRange;
+
+  ezUInt32 m_VertexBufferStrides[EZ_GAL_MAX_VERTEX_BUFFER_COUNT] = {};
+  ezUInt32 m_VertexBufferOffsets[EZ_GAL_MAX_VERTEX_BUFFER_COUNT] = {};
+};
