--- Code/Engine/RendererDX11/Context/Implementation/ContextDX11_inl.h
@@ -1,5 +1,0 @@
-
-ID3D11DeviceContext* ezGALContextDX11::GetDXContext() const
-{
-  return m_pDXContext;
-}
