+++ Code/Engine/RendererDX11/Device/DeviceDX11.h
@@ -2,6 +2,7 @@
 #pragma once
 
 #include <Foundation/Types/Bitflags.h>
+#include <Foundation/Types/UniquePtr.h>
 #include <RendererDX11/RendererDX11DLL.h>
 #include <RendererFoundation/Device/Device.h>
 
@@ -23,21 +24,22 @@ struct IDXGIAdapter;
 typedef ezGALFormatLookupEntry<DXGI_FORMAT, (DXGI_FORMAT)0> ezGALFormatLookupEntryDX11;
 typedef ezGALFormatLookupTable<ezGALFormatLookupEntryDX11> ezGALFormatLookupTableDX11;
 
+class ezGALPassDX11;
 
 /// \brief The DX11 device implementation of the graphics abstraction layer.
 class EZ_RENDERERDX11_DLL ezGALDeviceDX11 : public ezGALDevice
 {
-
-  /// \todo This shouldn't be accessible, there should be a factory instantiating the correct renderer class via RTTI for example
-public:
+private:
+  friend ezGALDevice* CreateDX11Device(ezAllocatorBase* pAllocator, const ezGALDeviceCreationDescription& Description);
   ezGALDeviceDX11(const ezGALDeviceCreationDescription& Description);
 
+public:
   virtual ~ezGALDeviceDX11();
-
 
 public:
   ID3D11Device* GetDXDevice() const;
   ID3D11Device3* GetDXDevice3() const;
+  ID3D11DeviceContext* GetDXImmediateContext() const;
 
   IDXGIFactory1* GetDXGIFactory() const;
 
@@ -56,77 +58,66 @@ protected:
   ezResult InitPlatform(DWORD flags, IDXGIAdapter* pUsedAdapter);
 
   virtual ezResult InitPlatform() override;
-
   virtual ezResult ShutdownPlatform() override;
+
+  // Pass functions
+
+  virtual ezGALPass* BeginPassPlatform(const char* szName) override;
+  virtual void EndPassPlatform(ezGALPass* pPass) override;
 
 
   // State creation functions
 
   virtual ezGALBlendState* CreateBlendStatePlatform(const ezGALBlendStateCreationDescription& Description) override;
-
   virtual void DestroyBlendStatePlatform(ezGALBlendState* pBlendState) override;
 
   virtual ezGALDepthStencilState* CreateDepthStencilStatePlatform(const ezGALDepthStencilStateCreationDescription& Description) override;
-
   virtual void DestroyDepthStencilStatePlatform(ezGALDepthStencilState* pDepthStencilState) override;
 
   virtual ezGALRasterizerState* CreateRasterizerStatePlatform(const ezGALRasterizerStateCreationDescription& Description) override;
-
   virtual void DestroyRasterizerStatePlatform(ezGALRasterizerState* pRasterizerState) override;
 
   virtual ezGALSamplerState* CreateSamplerStatePlatform(const ezGALSamplerStateCreationDescription& Description) override;
-
   virtual void DestroySamplerStatePlatform(ezGALSamplerState* pSamplerState) override;
 
 
   // Resource creation functions
 
   virtual ezGALShader* CreateShaderPlatform(const ezGALShaderCreationDescription& Description) override;
-
   virtual void DestroyShaderPlatform(ezGALShader* pShader) override;
 
   virtual ezGALBuffer* CreateBufferPlatform(const ezGALBufferCreationDescription& Description, ezArrayPtr<const ezUInt8> pInitialData) override;
-
   virtual void DestroyBufferPlatform(ezGALBuffer* pBuffer) override;
 
   virtual ezGALTexture* CreateTexturePlatform(const ezGALTextureCreationDescription& Description, ezArrayPtr<ezGALSystemMemoryDescription> pInitialData) override;
-
   virtual void DestroyTexturePlatform(ezGALTexture* pTexture) override;
 
   virtual ezGALResourceView* CreateResourceViewPlatform(ezGALResourceBase* pResource, const ezGALResourceViewCreationDescription& Description) override;
-
   virtual void DestroyResourceViewPlatform(ezGALResourceView* pResourceView) override;
 
   virtual ezGALRenderTargetView* CreateRenderTargetViewPlatform(ezGALTexture* pTexture, const ezGALRenderTargetViewCreationDescription& Description) override;
-
   virtual void DestroyRenderTargetViewPlatform(ezGALRenderTargetView* pRenderTargetView) override;
 
   ezGALUnorderedAccessView* CreateUnorderedAccessViewPlatform(ezGALResourceBase* pResource, const ezGALUnorderedAccessViewCreationDescription& Description) override;
-
   virtual void DestroyUnorderedAccessViewPlatform(ezGALUnorderedAccessView* pResource) override;
 
   // Other rendering creation functions
 
   virtual ezGALSwapChain* CreateSwapChainPlatform(const ezGALSwapChainCreationDescription& Description) override;
-
   virtual void DestroySwapChainPlatform(ezGALSwapChain* pSwapChain) override;
 
   virtual ezGALFence* CreateFencePlatform() override;
-
   virtual void DestroyFencePlatform(ezGALFence* pFence) override;
 
   virtual ezGALQuery* CreateQueryPlatform(const ezGALQueryCreationDescription& Description) override;
-
   virtual void DestroyQueryPlatform(ezGALQuery* pQuery) override;
 
   virtual ezGALVertexDeclaration* CreateVertexDeclarationPlatform(const ezGALVertexDeclarationCreationDescription& Description) override;
-
   virtual void DestroyVertexDeclarationPlatform(ezGALVertexDeclaration* pVertexDeclaration) override;
 
   // Timestamp functions
 
   virtual ezGALTimestampHandle GetTimestampPlatform() override;
-
   virtual ezResult GetTimestampResultPlatform(ezGALTimestampHandle hTimestamp, ezTime& result) override;
 
   // Swap chain functions
@@ -136,7 +127,6 @@ protected:
   // Misc functions
 
   virtual void BeginFramePlatform() override;
-
   virtual void EndFramePlatform() override;
 
   virtual void SetPrimarySwapChainPlatform(ezGALSwapChain* pSwapChain) override;
@@ -145,8 +135,14 @@ protected:
 
   /// \endcond
 
+public:
+  ID3D11Resource* FindTempBuffer(ezUInt32 uiSize);
+  ID3D11Resource* FindTempTexture(ezUInt32 uiWidth, ezUInt32 uiHeight, ezUInt32 uiDepth, ezGALResourceFormat::Enum format);
+
+  ID3D11Query* GetTimestamp(ezGALTimestampHandle hTimestamp);
+
 private:
-  friend class ezGALContextDX11;
+  friend class ezGALRenderCommandEncoderDX11;
 
   struct TempResourceType
   {
@@ -159,16 +155,13 @@ private:
     };
   };
 
-  ID3D11Resource* FindTempBuffer(ezUInt32 uiSize);
-  ID3D11Resource* FindTempTexture(ezUInt32 uiWidth, ezUInt32 uiHeight, ezUInt32 uiDepth, ezGALResourceFormat::Enum format);
   void FreeTempResources(ezUInt64 uiFrame);
-
-  ID3D11Query* GetTimestamp(ezGALTimestampHandle hTimestamp);
 
   void FillFormatLookupTable();
 
   ID3D11Device* m_pDevice;
   ID3D11Device3* m_pDevice3;
+  ID3D11DeviceContext* m_pImmediateContext;
 
   ID3D11Debug* m_pDebug;
 
@@ -181,6 +174,8 @@ private:
   ezGALFormatLookupTableDX11 m_FormatLookupTable;
 
   ezUInt32 m_FeatureLevel; // D3D_FEATURE_LEVEL can't be forward declared
+
+  ezUniquePtr<ezGALPassDX11> m_pDefaultPass;
 
   struct PerFrameData
   {
