+++ Code/Engine/RendererDX11/Resources/BufferDX11.h
@@ -9,9 +9,9 @@ struct ID3D11Buffer;
 class EZ_RENDERERDX11_DLL ezGALBufferDX11 : public ezGALBuffer
 {
 public:
-  EZ_ALWAYS_INLINE ID3D11Buffer* GetDXBuffer() const;
+  ID3D11Buffer* GetDXBuffer() const;
 
-  EZ_ALWAYS_INLINE DXGI_FORMAT GetIndexFormat() const;
+  DXGI_FORMAT GetIndexFormat() const;
 
 protected:
   friend class ezGALDeviceDX11;
