+++ Code/Engine/RendererDX11/Resources/Implementation/BufferDX11_inl.h
@@ -1,10 +1,10 @@
-﻿
-ID3D11Buffer* ezGALBufferDX11::GetDXBuffer() const
+
+EZ_ALWAYS_INLINE ID3D11Buffer* ezGALBufferDX11::GetDXBuffer() const
 {
   return m_pDXBuffer;
 }
 
-DXGI_FORMAT ezGALBufferDX11::GetIndexFormat() const
+EZ_ALWAYS_INLINE DXGI_FORMAT ezGALBufferDX11::GetIndexFormat() const
 {
   return m_IndexFormat;
 }
