+++ Code/Engine/RendererFoundation/Device/Implementation/DeviceFactory.cpp
@@ -13,10 +13,19 @@ static ezHashTable<ezString, CreatorFuncInfo> s_CreatorFuncs;
 
 CreatorFuncInfo* GetCreatorFuncInfo(const char* szRendererName)
 {
-  ezStringBuilder sb = szRendererName;
-  sb.ToLower();
+  auto pFuncInfo = s_CreatorFuncs.GetValue(szRendererName);
+  if (pFuncInfo == nullptr)
+  {
+    ezStringBuilder sPluginName = "ezRenderer";
+    sPluginName.Append(szRendererName);
 
-  return s_CreatorFuncs.GetValue(sb);
+    EZ_VERIFY(ezPlugin::LoadPlugin(sPluginName).Succeeded(), "Renderer plugin '{}' not found", sPluginName);
+
+    pFuncInfo = s_CreatorFuncs.GetValue(szRendererName);
+    EZ_ASSERT_DEV(pFuncInfo != nullptr, "Renderer '{}' is not registered", szRendererName);
+  }
+
+  return pFuncInfo;
 }
 
 ezGALDevice* ezGALDeviceFactory::CreateDevice(const char* szRendererName, ezAllocatorBase* pAllocator, const ezGALDeviceCreationDescription& desc)
@@ -36,4 +45,14 @@ void ezGALDeviceFactory::GetShaderModelAndCompiler(const char* szRendererName, c
     szShaderModel = pFuncInfo->m_sShaderModel;
     szShaderCompiler = pFuncInfo->m_sShaderCompiler;
   }
+}
+
+bool ezGALDeviceFactory::RegisterCreatorFunc(const char* szRendererName, const CreatorFunc& func, const char* szShaderModel, const char* szShaderCompiler)
+{
+  CreatorFuncInfo funcInfo;
+  funcInfo.m_Func = func;
+  funcInfo.m_sShaderModel = szShaderModel;
+  funcInfo.m_sShaderCompiler = szShaderCompiler;
+
+  return s_CreatorFuncs.Insert(szRendererName, funcInfo);
 }
